fix: music genres selection UI
