+++ app/src/main/java/com/epfl/beatlink/ui/components/profile/ProfileColumn.kt
@@ -17,6 +17,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
@@ -169,15 +170,14 @@ fun ProfileColumn(
         // Favorite music genres
         if (profileData?.favoriteMusicGenres?.isNotEmpty() == true) {
           GradientTitle("MUSIC GENRES")
-          Row(
+          LazyRow(
               modifier = Modifier.padding(vertical = 16.dp).testTag("favoriteMusicGenresRow"),
-              horizontalArrangement = Arrangement.spacedBy(16.dp),
-          ) {
-            profileData.favoriteMusicGenres.forEach { genre ->
-              val genreGradient = genreGradients[genre] ?: PrimaryGradientBrush
-              MusicGenreCard(genre = genre, brush = genreGradient, onClick = {})
-            }
-          }
+              horizontalArrangement = Arrangement.spacedBy(16.dp)) {
+                items(profileData.favoriteMusicGenres) { genre ->
+                  val genreGradient = genreGradients[genre] ?: PrimaryGradientBrush
+                  MusicGenreCard(genre = genre, brush = genreGradient, onClick = {})
+                }
+              }
         }
 
         // Display top songs if available
